Rearrange auth endpoint and test logic
