+++ endpoints/instantiations/auth_/resend_validation_email/endpoint_schema_config.py
@@ -1,6 +1,6 @@
+from endpoints.instantiations.auth_.resend_validation_email.dto import EmailOnlyDTO
+from endpoints.instantiations.auth_.resend_validation_email.schema import EmailOnlySchema
 from endpoints.schema_config.helpers import schema_config_with_message_output
-from middleware.schema_and_dto.dtos.signup import EmailOnlyDTO
-from middleware.schema_and_dto.schemas.signup import EmailOnlySchema
 
 AuthResendValidationEmailEndpointSchemaConfig = schema_config_with_message_output(
     input_schema=EmailOnlySchema(),
