+++ endpoints/instantiations/auth_/resend_validation_email/schema.py
@@ -1,4 +1,4 @@
-from middleware.schema_and_dto.dtos.signup import EmailOnlyDTO
+from endpoints.instantiations.auth_.resend_validation_email.dto import EmailOnlyDTO
 from middleware.schema_and_dto.dynamic.pydantic_to_marshmallow.core import (
     pydantic_to_marshmallow,
 )
