+++ endpoints/instantiations/auth_/validate_email/endpoint_schema_config.py
