+++ tests/integration/auth/signup/constants.py
@@ -1,0 +1,7 @@
+from db.enums import UserCapacityEnum
+
+PATCH_ROOT = "endpoints.instantiations.auth_.signup.middleware"
+TEST_USER_CAPACITIES = [
+    UserCapacityEnum.POLICE,
+    UserCapacityEnum.COMMUNITY_MEMBER,
+ ]
